requirements.txt
jsonschema==4.6.0
PyLD==2.0.3
pytest==7.1.2
PyYAML==6.0
typing_extensions==4.2.0
